requirements.txt
# Core API
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.18

# Validation & Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# LLM
openai==1.51.0
groq==0.11.0


# Embeddings & Vector Search
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4

# Document Parsing
pypdf==5.0.0
python-docx==1.1.2

# Caching
cachetools==5.5.0

# HTTP
httpx==0.27.2

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Starlette pinned — avoids 1.0.0 which breaks things
starlette==0.38.6